+++ src/nifgen/metadata/functions_addon.py
@@ -60,6 +60,7 @@ functions_codegen_method = {
     'revision_query':                       { 'codegen_method': 'no',       },
     '.+Complex.+':                          { 'codegen_method': 'no',       },
     'GetStreamEndpointHandle':              { 'codegen_method': 'no',       },
+    'GetFIRFilterCoefficients':             { 'codegen_method': 'no',       },  # Removed - applies to OSP only #596 - If this is removed, the commented out snippet below needs to be added back to templates to use
     'AdjustSampleClockRelativeDelay':       { 'codegen_method': 'no',       },  # This is used internally by NI-TClk, but not by end users.
     '.etAttributeViInt64':                  { 'codegen_method': 'no',       },  # NI-FGEN has no ViInt64 attributes.
     'GetExtCalLastDateAndTime':             { 'codegen_method': 'private', 'method_name_for_documentation': 'get_ext_cal_last_date_and_time',  },  # 'GetLastExtCalLastDateAndTime' Public wrapper to allow datetime
@@ -98,12 +99,11 @@ functions_enums = {
     'SetNamedWaveformNextWritePosition':        { 'parameters': { 3: { 'enum': 'RelativeTo',                }, }, },
     'SetWaveformNextWritePosition':             { 'parameters': { 3: { 'enum': 'RelativeTo',                }, }, },
     'GetHardwareState':                         { 'parameters': { 1: { 'enum': 'HardwareState',             }, }, },
-    'SendSoftwareEdgeTrigger':                  { 'parameters': { 1: { 'enum': 'Trigger',                   }, }, },  # TODO: issue #538
 }
 
-functions_issues = {
-    'GetFIRFilterCoefficients':             { 'parameters': { 3: { 'direction':'out'},  # TODO(marcoskirsch): Remove when #534 solved
-                                                              4: { 'direction':'out', 'is_buffer': False, 'type':'ViInt32', }, }, },
+functions_send_software_edge_trigger = {
+    'SendSoftwareEdgeTrigger':                  { 'method_templates': [ { 'session_filename': 'send_software_edge_trigger', 'documentation_filename': 'send_software_edge_trigger', 'method_python_name_suffix': '', }, ],
+                                                  'render_in_session_base': True, },
 }
 
 # This is the additional metadata needed by the code generator in order create code that can properly handle buffer allocation.
@@ -116,7 +116,6 @@ functions_buffer_info = {
     'ConfigureCustomFIRFilterCoefficients': { 'parameters': { 3: { 'size': {'mechanism':'len', 'value':'numberOfCoefficients'}, }, }, },
     'CreateWaveform(I16|F64)':              { 'parameters': { 3: { 'size': {'mechanism':'len', 'value':'waveformSize'}, }, }, },
     'DefineUserStandardWaveform':           { 'parameters': { 3: { 'size': {'mechanism':'len', 'value':'waveformSize'}, }, }, },
-    'GetFIRFilterCoefficients':             { 'parameters': { 3: { 'size': {'mechanism':'ivi-dance', 'value':'arraySize'}, }, }, },  # TODO(marcoskirsch): #537
     'Write.*Waveform':                      { 'parameters': { 4: { 'size': {'mechanism':'len', 'value':'Size'}, }, }, },
     'CreateAdvancedArbSequence':            { 'parameters': { 2: { 'size': {'mechanism':'len', 'value':'sequenceLength'}, },
                                                               3: { 'size': {'mechanism':'len', 'value':'sequenceLength'}, },
@@ -502,6 +501,11 @@ functions_method_templates = {
         { 'session_filename': 'numpy_write_method', 'method_python_name_suffix': '_numpy', },
     ], },
 }
+
+# We keep this information because we will need it again if we ever enable OSP and need this function
+# 'GetFIRFilterCoefficients':     { 'method_templates': [
+#     { 'session_filename': 'get_fir_filter_coefficients', 'documentation_filename': 'get_fir_filter_coefficients', 'method_python_name_suffix': '', },
+# ], },
 
 functions_numpy = {
     'CreateWaveformF64':            { 'parameters': { 3: { 'numpy': True, }, }, },