+++ src/niscope/system_tests/test_system_niscope.py
@@ -3,6 +3,7 @@ import niscope
 import numpy
 import pytest
 import sys
+import tempfile
 
 
 @pytest.fixture(scope='function')
@@ -93,11 +94,7 @@ def test_fetch_binary8_into(session):
             for j in range(len(record_wfm)):
                 assert record_wfm[j] == waveform[i * test_record_length + j]
         else:
-            try:
-                waveforms[i].wfm
-                assert False
-            except AttributeError:
-                pass
+            assert waveforms[i].samples is None
 
 
 def test_fetch_binary16_into(session):
@@ -125,11 +122,7 @@ def test_fetch_binary16_into(session):
             for j in range(len(record_wfm)):
                 assert record_wfm[j] == waveform[i * test_record_length + j]
         else:
-            try:
-                waveforms[i].samples
-                assert False
-            except AttributeError:
-                pass
+            assert waveforms[i].samples is None
 
 
 def test_fetch_binary32_into(session):
@@ -157,11 +150,7 @@ def test_fetch_binary32_into(session):
             for j in range(len(record_wfm)):
                 assert record_wfm[j] == waveform[i * test_record_length + j]
         else:
-            try:
-                waveforms[i].samples
-                assert False
-            except AttributeError:
-                pass
+            assert waveforms[i].samples is None
 
 
 def test_fetch_double_into(session):
@@ -189,11 +178,7 @@ def test_fetch_double_into(session):
             for j in range(len(record_wfm)):
                 assert record_wfm[j] == waveform[i * test_record_length + j]
         else:
-            try:
-                waveforms[i].samples
-                assert False
-            except AttributeError:
-                pass
+            assert waveforms[i].samples is None
 
 
 def test_self_test(session):
@@ -326,6 +311,31 @@ def test_configure_trigger_hysteresis(session):
     session.configure_trigger_hysteresis('1', 0.0, 0.05, niscope.TriggerCoupling.DC)
     assert '1' == session.trigger_source
     assert niscope.TriggerCoupling.DC == session.trigger_coupling
+
+
+def test_import_export_buffer(session):
+    test_value_1 = 1
+    test_value_2 = 5
+    session.vertical_range = test_value_1
+    assert session.vertical_range == test_value_1
+    buffer = session.export_attribute_configuration_buffer()
+    session.vertical_range = test_value_2
+    assert session.vertical_range == test_value_2
+    session.import_attribute_configuration_buffer(buffer)
+    assert session.vertical_range == test_value_1
+
+
+def test_import_export_file(session):
+    test_value_1 = 1
+    test_value_2 = 5
+    path = tempfile.gettempdir() + 'test.txt'
+    session.vertical_range = test_value_1
+    assert session.vertical_range == test_value_1
+    session.export_attribute_configuration_file(path)
+    session.vertical_range = test_value_2
+    assert session.vertical_range == test_value_2
+    session.import_attribute_configuration_file(path)
+    assert session.vertical_range == test_value_1
 
 
 def test_configure_trigger_software(session):