+++ src/niswitch/metadata/config.py
@@ -1,7 +1,7 @@
 config = {
     'metadata_version': '1.0',
     'module_name': 'niswitch',
-    'module_version': '1.0.0.dev0',
+    'module_version': '1.0.2.dev0',
     'c_function_prefix': 'niSwitch_',
     'driver_name': 'NI-SWITCH',
     'session_class_description': 'An NI-SWITCH session to a National Instruments Switch Module',
@@ -22,8 +22,9 @@ config = {
         'abort_function': 'AbortScan',
     },
     'init_function': 'InitWithTopology',
+    'close_function': 'close',
     'custom_types': [],
-    'last_tested_version': '17.0.0',
+    'last_tested_version': '18.0.0',
     'repeated_capabilities': [
         {'python_name': 'channels', 'prefix': '', },
     ],